requirements.txt
# Regular dependencies
textual==6.2.0
requests==2.31.0
pyserial==3.5
skyfield==1.48
fastapi[standard]==0.115.0
fastapi-cache2==0.2.1
sqlmodel==0.0.22
jinja2==3.1.6
toml==0.10.2
loguru==0.7.3
pyham_ax25==1.0.2
tinyaes==1.1.1
pyStuffing==0.0.4
hypothesis==6.131.30
psycopg2-binary==2.9.10
python-dotenv==1.1.0
tqdm==4.67
uvicorn==0.22.0
pydantic-settings==2.4.0

# Typed packages
types-requests==2.31.0
types-pyserial==3.5.0
types-setuptools==69.1.0.20240217
types-toml==0.10.8
types-tqdm==4.67.0.20250516

# Development dependencies
alembic==1.11.1
pre-commit==3.3.3
pytest==7.4.0
pytest-cov==4.1.0
mypy==1.8.0
ruff==0.2.0
psycopg[binary]
pytest-postgresql==7.0.1
